Fix compile errors in governance, tests, and registration guards.

+++ contracts/vault/src/permissions.rs
@@ -59,10 +59,10 @@ impl MultiSignerValidator {
         if threshold == 0 {
             return Err("threshold must be > 0");
         }
-        if threshold as usize > signers.len() {
+        if threshold > signers.len() as u32 {
             return Err("threshold exceeds signer set size");
         }
-        if approvals.len() < threshold as usize {
+        if approvals.len() as u32 < threshold {
             return Err("insufficient approvals");
         }
